requirements.txt
streamlit
pdfplumber
python-docx
openai
scikit-learn
numpy
pandas
tqdm